owner based auth added

+++ PlateRate.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,11 +2,14 @@
 using Microsoft.Extensions.Logging;
 using PlateRate.Domain.Entities;
 using PlateRate.Domain.Exceptions;
+using PlateRate.Domain.Interfaces;
 using PlateRate.Domain.Repositories;
+using PlateRate.Infrastructure.Constants;
 
 namespace PlateRate.Application.Restaurants.Commands.DeleteRestaurant;
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger
-    ,IRestaurantRepository restaurantRepository) : IRequestHandler<DeleteRestaurantCommand>
+    ,IRestaurantRepository restaurantRepository
+    ,IRestaurantAuthorizationService restaurantAuthorization) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -16,6 +19,11 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         {
             throw new NotFoundException(nameof(Restaurant),request.Id.ToString());
         }
+        if (!restaurantAuthorization.Authorize(restaurant,ResourceOperation.Delete))
+        {
+            throw new ForbidException();
+        }
+
         await restaurantRepository.DeleteAsync(restaurant);
     }
 }
